delete functionality in YearlyChart and add everything to reducer
