+++ workout-tracker/src/store/reducers/charts.js
@@ -234,6 +234,99 @@ const charts = (state = initialState, action) => {
         monthlyChart: CopyOfMonthlyChart
       };
 
+
+
+
+      let workoutNames = [];
+    let workouts = [];
+
+    this.props.workouts.map(workout => {
+      workoutNames.push(workout.workout_name);
+      workouts.push(workout);
+      return workout;
+    });
+
+    let year = new Date().getFullYear();
+    let first_day_year = new Date(year, 0, 1);
+    let last_day_year = new Date(year, 11, 31);
+
+    let getDaysArray = function(s, e) {
+      for (var a = [], d = s; d <= e; d.setDate(d.getDate() + 1)) {
+        a.push(new Date(d));
+      }
+      return a;
+    };
+
+    let daylist = getDaysArray(first_day_year, last_day_year);
+    daylist.map(v => v.toISOString().slice(0, 10)).join("");
+
+    let daysInYear = [];
+
+    function formatDate(date) {
+      let d = new Date(date),
+        month = "" + (d.getMonth() + 1),
+        day = "" + d.getDate(),
+        year = d.getFullYear();
+
+      if (month.length < 2) month = "0" + month;
+      if (day.length < 2) day = "0" + day;
+
+      return [year, month, day].join("-");
+    }
+
+    for (let i = 0; i < daylist.length; i++) {
+      daysInYear.push(
+        formatDate(daylist[i])
+          .split("-")
+          .join("")
+      );
+    }
+
+    let userHistory = this.props.history;
+    let resultOfWeek = [];
+
+    for (let j = 0; j < daysInYear.length; j++) {
+      for (let i = 0; i < userHistory.length; i++) {
+        if (
+          userHistory[i].session_start
+            .match(/.{1,10}/g)[0]
+            .split("-")
+            .join("") === daysInYear[j]
+        ) {
+          resultOfWeek.push(userHistory[i]);
+        }
+      }
+    }
+
+    let hashTable = {};
+
+    for (let j = 0; j < workouts.length; j++) {
+      hashTable[workouts[j].workout_name] = 0;
+    }
+
+    for (let i = 0; i < resultOfWeek.length; i++) {
+      for (let j = 0; j < workouts.length; j++) {
+        if (resultOfWeek[i].workout_id === workouts[j].id) {
+          if (hashTable[workouts[j].workout_name]) {
+            hashTable[workouts[j].workout_name] += 1;
+          } else {
+            hashTable[workouts[j].workout_name] = 1;
+          }
+        }
+      }
+    }
+
+    let valuesForDataset = [];
+
+    for (var value in hashTable) {
+      valuesForDataset.push(hashTable[value]);
+    }
+
+    this.setState({
+      data: valuesForDataset,
+      labels: workoutNames
+    });
+
     default:
       return state;
   }
