+++ workout-tracker/src/views/Stats/Charts/YearlyChart.js
@@ -2,127 +2,13 @@ import React from "react";
 import { Pie } from "react-chartjs-2";
 import { Card } from "antd";
 import { connect } from "react-redux";
+import { calculateYearlyChart } from "../../../store/actions/chartActions";
 
 const { Meta } = Card;
 
 class YearlyChart extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      labels: ["Red", "Green", "Yellow"],
-      data: [],
-      backgroundColor: [
-        "#f6f078",
-        "#01d28e",
-        "#434982",
-        "#730068",
-        "#a6e3e9",
-        "##36A2EB",
-        "#51dacf",
-        "#edaaaa"
-      ],
-      hoverBackgroundColor: [
-        "#f6f078",
-        "#01d28e",
-        "#434982",
-        "#730068",
-        "#a6e3e9",
-        "##36A2EB",
-        "#51dacf",
-        "#edaaaa"
-      ]
-    };
-  }
   componentDidMount = () => {
-    let workoutNames = [];
-    let workouts = [];
-
-    this.props.workouts.map(workout => {
-      workoutNames.push(workout.workout_name);
-      workouts.push(workout);
-      return workout;
-    });
-
-    let year = new Date().getFullYear();
-    let first_day_year = new Date(year, 0, 1);
-    let last_day_year = new Date(year, 11, 31);
-
-    let getDaysArray = function(s, e) {
-      for (var a = [], d = s; d <= e; d.setDate(d.getDate() + 1)) {
-        a.push(new Date(d));
-      }
-      return a;
-    };
-
-    let daylist = getDaysArray(first_day_year, last_day_year);
-    daylist.map(v => v.toISOString().slice(0, 10)).join("");
-
-    let daysInYear = [];
-
-    function formatDate(date) {
-      let d = new Date(date),
-        month = "" + (d.getMonth() + 1),
-        day = "" + d.getDate(),
-        year = d.getFullYear();
-
-      if (month.length < 2) month = "0" + month;
-      if (day.length < 2) day = "0" + day;
-
-      return [year, month, day].join("-");
-    }
-
-    for (let i = 0; i < daylist.length; i++) {
-      daysInYear.push(
-        formatDate(daylist[i])
-          .split("-")
-          .join("")
-      );
-    }
-
-    let userHistory = this.props.history;
-    let resultOfWeek = [];
-
-    for (let j = 0; j < daysInYear.length; j++) {
-      for (let i = 0; i < userHistory.length; i++) {
-        if (
-          userHistory[i].session_start
-            .match(/.{1,10}/g)[0]
-            .split("-")
-            .join("") === daysInYear[j]
-        ) {
-          resultOfWeek.push(userHistory[i]);
-        }
-      }
-    }
-
-    let hashTable = {};
-
-    for (let j = 0; j < workouts.length; j++) {
-      hashTable[workouts[j].workout_name] = 0;
-    }
-
-    for (let i = 0; i < resultOfWeek.length; i++) {
-      for (let j = 0; j < workouts.length; j++) {
-        if (resultOfWeek[i].workout_id === workouts[j].id) {
-          if (hashTable[workouts[j].workout_name]) {
-            hashTable[workouts[j].workout_name] += 1;
-          } else {
-            hashTable[workouts[j].workout_name] = 1;
-          }
-        }
-      }
-    }
-
-    let valuesForDataset = [];
-
-    for (var value in hashTable) {
-      valuesForDataset.push(hashTable[value]);
-    }
-
-    this.setState({
-      data: valuesForDataset,
-      labels: workoutNames
-    });
+    this.props.calculateYearlyChart(this.props.history, this.props.workouts);
   };
 
   render() {
@@ -141,14 +27,15 @@ class YearlyChart extends React.Component {
               borderTopRightRadius: ".6rem",
             }}
           >
-            <Pie
+             <Pie
               data={{
-                labels: this.state.labels,
+                labels: this.props.yearlyChart.labels,
                 datasets: [
                   {
-                    data: this.state.data,
-                    backgroundColor: this.state.backgroundColor,
-                    hoverBackgroundColor: this.state.hoverBackgroundColor
+                    data: this.props.yearlyChart.data,
+                    backgroundColor: this.props.yearlyChart.backgroundColor,
+                    hoverBackgroundColor: this.props.yearlyChart
+                      .hoverBackgroundColor
                   }
                 ]
               }}
@@ -161,11 +48,11 @@ class YearlyChart extends React.Component {
           description={
             <div>
               <i className="fa fa-fire"></i>{" "}
-              {` You made ${this.state.data.reduce(
+              {` You made ${this.props.yearlyChart.data.reduce(
                 (accumulator, currentValue) => accumulator + currentValue,
                 0
               )} ${
-                this.state.data.reduce(
+                this.state.props.yearlyChart.reduce(
                   (accumulator, currentValue) => accumulator + currentValue,
                   0
                 ) === 1
@@ -183,8 +70,9 @@ class YearlyChart extends React.Component {
 const mapStateToProps = state => {
   return {
     history: state.history.history,
-    workouts: state.workouts.workouts
+    workouts: state.workouts.workouts,
+    yearlyChart: state.charts.yearlyChart
   };
 };
 
-export default connect(mapStateToProps)(YearlyChart);
+export default connect(mapStateToProps, { calculateYearlyChart })(YearlyChart);
